Tower Defender: Part 2 (#2)
* Imported Recorder

* Turret shop Setuo

Player can choose to place between 2 turrets by clicking it's associated button

* Set up Missile Launcher prefab

* Player can place Missile Launcher objects on board

* Implemented missile functionality

If the missile collides with its target, it damages any enemy in its range.

* Created a Particle effect for missile impact

* Implemented Currency system

Placing any turret now costs a unique amount of currency. The player can only place a turret if they have enough currency.

* Implemented Currency UI

Currency is now displayed, and updated when the player purchases a turret

* Created a particle effect for placing a turret

* Set up the Lives UI

* Enemy Health implemented

Shooting an enemy now reduces their health, rather than immediately destroying them. If the enemy's health is zero or less, it'll be destroyed.

* Enemy Death particles implemented

* LaserBeam turret object setup

* Player can purchase a Laser Beamer turret

* Laser points at passing enemies

* Created Laser Contact particles

* Set placement of laser contact particles

* Layered Particle Effect for Laser Contact created

The Laser Particles now has a glow particle effect and a point light nested under it

* Damage Over Time for Laser Beamer implemented

* Laser slows enemys

An enemy's speed is now slowed while its being hit by a laser

* Game Over UI setup

* Restart functionality implemented

* Animated Game Over Screen pop up

* NodeUI setup

Created the visual for the Upgrade / Sell option that appears when any node with a turret is clicked on. This visual snaps over any turret that has been clicked, and is hidden when the player clicks away

* Turret upgrade functionality implemented

* Upgrade Button visual changes once upgraded